+++ cmd/cloud-server/service/account/sync.go
@@ -20,30 +20,11 @@
 package account
 
 import (
-	"errors"
-	"fmt"
-	"strings"
-
-	"hcm/cmd/cloud-server/service/sync/aws"
-	"hcm/cmd/cloud-server/service/sync/azure"
-	"hcm/cmd/cloud-server/service/sync/gcp"
-	"hcm/cmd/cloud-server/service/sync/huawei"
-	"hcm/cmd/cloud-server/service/sync/lock"
-	"hcm/cmd/cloud-server/service/sync/tcloud"
-	"hcm/pkg/api/core"
-	protoregion "hcm/pkg/api/data-service/cloud/region"
-	protocloud "hcm/pkg/api/data-service/cloud/zone"
-	dataservice "hcm/pkg/client/data-service"
-	"hcm/pkg/criteria/constant"
+	"hcm/cmd/cloud-server/logics/account"
 	"hcm/pkg/criteria/enumor"
 	"hcm/pkg/criteria/errf"
-	"hcm/pkg/dal/dao/tools"
 	"hcm/pkg/iam/meta"
-	"hcm/pkg/kit"
-	"hcm/pkg/logs"
 	"hcm/pkg/rest"
-
-	etcd3 "go.etcd.io/etcd/client/v3"
 )
 
 // SyncCloudResource ...
@@ -66,203 +47,9 @@ func (a *accountSvc) SyncCloudResource(cts *rest.Contexts) (interface{}, error) 
 		return nil, errf.NewFromErr(errf.InvalidParameter, err)
 	}
 
-	isNeedSyncPublicResFlag, err := isNeedSyncPublicResource(cts.Kit, a.client.DataService(), baseInfo.Vendor)
-	if err != nil {
-		logs.Errorf("is need sync public resource failed, err: %v, vendor: %s, rid: %s", err,
-			baseInfo.Vendor, cts.Kit.Rid)
+	if err = account.Sync(cts.Kit, a.client, baseInfo.Vendor, accountID); err != nil {
 		return nil, err
 	}
-
-	leaseID, err := lock.Manager.TryLock(lock.Key(accountID))
-	if err != nil {
-		if err == lock.ErrLockFailed {
-			return nil, errors.New("synchronization is in progress")
-		}
-
-		return nil, err
-	}
-
-	go func(leaseID etcd3.LeaseID) {
-		defer func() {
-			if err := lock.Manager.UnLock(leaseID); err != nil {
-				// 锁已经超时释放了
-				if strings.Contains(err.Error(), "requested lease not found") {
-					return
-				}
-
-				logs.Errorf("%s: unlock account sync lock failed, err: %v, accountID: %s, leaseID: %d, rid: %s",
-					constant.AccountSyncFailed, err, accountID, leaseID, cts.Kit.Rid)
-			}
-		}()
-
-		a.syncAllResource(cts.Kit, baseInfo.Vendor, accountID, isNeedSyncPublicResFlag)
-	}(leaseID)
 
 	return nil, nil
-}
-
-func (a *accountSvc) syncAllResource(kt *kit.Kit, vendor enumor.Vendor, accountID string, isNeed bool) {
-	switch vendor {
-	case enumor.TCloud:
-		opt := &tcloud.SyncAllResourceOption{
-			AccountID:          accountID,
-			SyncPublicResource: isNeed,
-		}
-		tcloud.SyncAllResource(kt, a.client, opt)
-
-	case enumor.Aws:
-		opt := &aws.SyncAllResourceOption{
-			AccountID:          accountID,
-			SyncPublicResource: isNeed,
-		}
-		aws.SyncAllResource(kt, a.client, opt)
-
-	case enumor.HuaWei:
-		opt := &huawei.SyncAllResourceOption{
-			AccountID:          accountID,
-			SyncPublicResource: isNeed,
-		}
-		huawei.SyncAllResource(kt, a.client, opt)
-
-	case enumor.Gcp:
-		opt := &gcp.SyncAllResourceOption{
-			AccountID:          accountID,
-			SyncPublicResource: isNeed,
-		}
-		gcp.SyncAllResource(kt, a.client, opt)
-
-	case enumor.Azure:
-		opt := &azure.SyncAllResourceOption{
-			AccountID:          accountID,
-			SyncPublicResource: isNeed,
-		}
-		azure.SyncAllResource(kt, a.client, opt)
-
-	default:
-		logs.Errorf("account: %s's vendor not support, vendor: %s", accountID, vendor)
-	}
-}
-
-func isNeedSyncPublicResource(kt *kit.Kit, dataCli *dataservice.Client, vendor enumor.Vendor) (
-		bool, error) {
-
-	need, err := isNeedSyncRegion(kt, dataCli, vendor)
-	if err != nil {
-		return false, err
-	}
-
-	if need {
-		return true, nil
-	}
-
-	switch vendor {
-	case enumor.Aws, enumor.TCloud, enumor.HuaWei, enumor.Gcp:
-		listZoneReq := &protocloud.ZoneListReq{
-			Filter: tools.EqualExpression("vendor", vendor),
-			Page:   core.NewCountPage(),
-		}
-		result, err := dataCli.Global.Zone.ListZone(kt.Ctx, kt.Header(), listZoneReq)
-		if err != nil {
-			return false, err
-		}
-
-		if result.Count == 0 {
-			return true, nil
-		}
-
-	case enumor.Azure:
-		// azure没有可用区
-	default:
-		return false, fmt.Errorf("vendor: %s not support", vendor)
-	}
-
-	switch vendor {
-	case enumor.Aws, enumor.TCloud, enumor.HuaWei, enumor.Gcp, enumor.Azure:
-		listZoneReq := &core.ListReq{
-			Filter: tools.EqualExpression("vendor", vendor),
-			Page:   core.NewCountPage(),
-		}
-		result, err := dataCli.Global.ListImage(kt, listZoneReq)
-		if err != nil {
-			return false, err
-		}
-
-		if result.Count == 0 {
-			return true, nil
-		}
-
-	default:
-		return false, fmt.Errorf("vendor: %s not support", vendor)
-	}
-
-	return false, nil
-}
-
-func isNeedSyncRegion(kt *kit.Kit, dataCli *dataservice.Client, vendor enumor.Vendor) (bool, error) {
-	regionCount := uint64(0)
-	switch vendor {
-	case enumor.TCloud:
-		listReq := &protoregion.TCloudRegionListReq{
-			Filter: tools.AllExpression(),
-			Page:   core.NewCountPage(),
-		}
-		result, err := dataCli.TCloud.Region.ListRegion(kt.Ctx, kt.Header(), listReq)
-		if err != nil {
-			return false, err
-		}
-		regionCount = result.Count
-
-	case enumor.Aws:
-		listReq := &protoregion.AwsRegionListReq{
-			Filter: tools.AllExpression(),
-			Page:   core.NewCountPage(),
-		}
-		result, err := dataCli.Aws.Region.ListRegion(kt.Ctx, kt.Header(), listReq)
-		if err != nil {
-			return false, err
-		}
-		regionCount = result.Count
-
-	case enumor.HuaWei:
-		listReq := &protoregion.HuaWeiRegionListReq{
-			Filter: tools.AllExpression(),
-			Page:   core.NewCountPage(),
-		}
-		result, err := dataCli.HuaWei.Region.ListRegion(kt.Ctx, kt.Header(), listReq)
-		if err != nil {
-			return false, err
-		}
-		regionCount = result.Count
-
-	case enumor.Gcp:
-		listReq := &protoregion.GcpRegionListReq{
-			Filter: tools.AllExpression(),
-			Page:   core.NewCountPage(),
-		}
-		result, err := dataCli.Gcp.Region.ListRegion(kt.Ctx, kt.Header(), listReq)
-		if err != nil {
-			return false, err
-		}
-		regionCount = result.Count
-
-	case enumor.Azure:
-		listReq := &protoregion.AzureRegionListReq{
-			Filter: tools.AllExpression(),
-			Page:   core.NewCountPage(),
-		}
-		result, err := dataCli.Azure.Region.ListRegion(kt.Ctx, kt.Header(), listReq)
-		if err != nil {
-			return false, err
-		}
-		regionCount = result.Count
-
-	default:
-		return false, fmt.Errorf("vendor: %s not support", vendor)
-	}
-
-	if regionCount == 0 {
-		return true, nil
-	}
-
-	return false, nil
 }
